Alterações de atributos

+++ src/main/java/br/com/fecorp/reinf/model/ComplementoMestreEntrada.java
@@ -1,6 +1,7 @@
 package br.com.fecorp.reinf.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ViewScoped;
@@ -26,11 +27,11 @@ public class ComplementoMestreEntrada implements Serializable {
         private String tipoProcesso;
         private String numProcessoAdm;
         private String indicativoSusp;
-        private double valorRetencaoPrincipal;
+        private BigDecimal valorRetencaoPrincipal;
         private String tipoProcesso01;
         private String numProcessoAdm02;
         private String indicativoSusp02;
-        private double valorRetencaoAdicional;
+        private BigDecimal valorRetencaoAdicional;
         
 		public String getCodEmpresa() {
 			return codEmpresa;
@@ -62,10 +63,10 @@ public class ComplementoMestreEntrada implements Serializable {
 		public void setIndicativoSusp(String indicativoSusp) {
 			this.indicativoSusp = indicativoSusp;
 		}
-		public double getValorRetencaoPrincipal() {
+		public BigDecimal getValorRetencaoPrincipal() {
 			return valorRetencaoPrincipal;
 		}
-		public void setValorRetencaoPrincipal(double valorRetencaoPrincipal) {
+		public void setValorRetencaoPrincipal(BigDecimal valorRetencaoPrincipal) {
 			this.valorRetencaoPrincipal = valorRetencaoPrincipal;
 		}
 		public String getTipoProcesso01() {
@@ -86,10 +87,10 @@ public class ComplementoMestreEntrada implements Serializable {
 		public void setIndicativoSusp02(String indicativoSusp02) {
 			this.indicativoSusp02 = indicativoSusp02;
 		}
-		public double getValorRetencaoAdicional() {
+		public BigDecimal getValorRetencaoAdicional() {
 			return valorRetencaoAdicional;
 		}
-		public void setValorRetencaoAdicional(double valorRetencaoAdicional) {
+		public void setValorRetencaoAdicional(BigDecimal valorRetencaoAdicional) {
 			this.valorRetencaoAdicional = valorRetencaoAdicional;
 		}
 		public static long getSerialversionuid() {
